+++ NEWS.md
@@ -2,8 +2,9 @@
 ---------------------
 * Use feature-wise permutation for propr()
 * Fix bug when calling updateCutoffs with func
-* Add trend=TRUE flag for updateF
 * Use harmonic mean when calculating weighted lrv and lrm
+* Add trend=TRUE flag for updateF to use mean-variance trend in limma moderation
+* Use sample weights for limma when calculating weighted theta in propd()
 
 # propr 5.1.7
 ---------------------